split cliruntimefixture into .h and .cpp
