+++ tests/src/cliruntimefixture.h
@@ -5,49 +5,17 @@
 #include "lute/runtime.h"
 
 #include "lua.h"
-#include "lualib.h"
-
-#include "Luau/Compiler.h"
 
 #include <memory>
-
-static int capture(lua_State* L)
-{
-    const char* str = luaL_tolstring(L, 1, nullptr);
-    lua_pushstring(L, "capturedoutput");
-    lua_pushstring(L, str ? str : "");
-    lua_settable(L, LUA_REGISTRYINDEX);
-    return 0;
-}
 
 class CliRuntimeFixture
 {
 public:
-    CliRuntimeFixture()
-        : runtime(std::make_unique<Runtime>())
-    {
-        L = setupCliState(*runtime, [](lua_State* L) {
-            lua_pushstring(L, "capturedoutput");
-            lua_pushstring(L, "");
-            lua_settable(L, LUA_REGISTRYINDEX);
-            lua_pushcfunction(L, capture, "capture");
-            lua_setglobal(L, "capture");
-        });
-    }
+    CliRuntimeFixture();
 
-    std::string getCapturedOutput()
-    {
-        lua_getfield(L, LUA_REGISTRYINDEX, "capturedoutput");
-        const char* output = lua_tostring(L, -1);
-        lua_pop(L, 1);
-        return output ? output : "";
-    }
+    std::string getCapturedOutput();
 
-    bool runCode(const std::string& source)
-    {
-        std::string bytecode = Luau::compile(source, Luau::CompileOptions());
-        return runBytecode(*runtime, bytecode, "=stdin", L, 0, nullptr);
-    }
+    bool runCode(const std::string& source);
 
     lua_State* L;
